+++ src/main/java/io/jeidiiy/outflearn/security/login/ajax/domain/AjaxLoginFailureDto.java
@@ -1,4 +1,4 @@
-package io.jeidiiy.outflearn.security.login.ajax;
+package io.jeidiiy.outflearn.security.login.ajax.domain;
 
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
