+++ src/main/java/io/jeidiiy/outflearn/security/login/ajax/domain/EmailLoginRequest.java
@@ -1,4 +1,4 @@
-package io.jeidiiy.outflearn.security.login.ajax;
+package io.jeidiiy.outflearn.security.login.ajax.domain;
 
 import lombok.Builder;
 import lombok.Getter;
