+++ src/main/java/io/jeidiiy/outflearn/security/login/ajax/handler/AjaxAuthenticationSuccessHandler.java
@@ -1,4 +1,4 @@
-package io.jeidiiy.outflearn.security.login.ajax;
+package io.jeidiiy.outflearn.security.login.ajax.handler;
 
 import java.io.IOException;
 
@@ -12,6 +12,8 @@ import org.springframework.security.web.savedrequest.SavedRequest;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 
+import io.jeidiiy.outflearn.security.login.ajax.domain.AjaxLoginSuccessDto;
+import io.jeidiiy.outflearn.security.login.ajax.domain.LoginUser;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.extern.slf4j.Slf4j;
